# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CookiesStatusData:
    properties:
      file_path:
        type: string
      file_size:
        type: integer
      has_cookies:
        type: boolean
      update_time:
        type: string
    type: object
...
  handler.Response:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.UpdateUserInfoRequest:
    properties:
      avatar:
        type: string
      username:
        type: string
    type: object
...
  handler.UserInfoResponse:
    properties:
      display_name:
        description: 显示名称（Firebase：username 或 email 前缀；邮箱：email 前缀）
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: 用户唯一 ID（Firebase 路径为 firebase_uid，邮箱路径为 email-auth SDK 分配的 ID）
        type: string
      photo_url:
        description: 头像 URL，无则为空字符串
        type: string
      provider:
        description: 登录来源：firebase | email
        type: string
    type: object
...
host: 127.0.0.1:8096
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for downloading YouTube videos and uploading to Bilibili with AI-powered
    processing
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: YouTube to Bilibili API
  version: "1.0"
paths:
...
  /api/user/update:
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的信息
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            $ref: '#/definitions/handler.UserInfoResponse'
        "400":
          description: 'error: string'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: string'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - user
...
  /api/v1/cookies:
    delete:
      consumes:
      - application/json
      description: 删除已上传的 YouTube cookies 文件。删除后需要重新上传才能下载受限制的视频
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.CookiesStatusData'
              type: object
        "404":
          description: Cookies 文件不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 删除文件失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除 cookies 文件
      tags:
      - cookies
  /api/v1/cookies/status:
    get:
      consumes:
      - application/json
      description: 检查 cookies 文件是否存在，返回文件路径、大小和最后更新时间等详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回文件状态
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.CookiesStatusData'
              type: object
        "500":
          description: 检查文件失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 cookies 文件状态
      tags:
      - cookies
  /api/v1/cookies/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传 Netscape 格式的 cookies.txt 文件，用于 yt-dlp 下载受限制的 YouTube 视频。文件需从浏览器导出，最大支持
        10MB
      parameters:
      - description: Cookies 文件 (Netscape 格式，.txt 文件)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.CookiesStatusData'
              type: object
        "400":
          description: 参数错误：文件格式不正确、文件过大或文件格式无效
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器错误：文件写入失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 上传 YouTube cookies 文件
      tags:
      - cookies
...
  /api/v1/updater/check:
    post:
      consumes:
      - application/json
      description: 检查 GitHub Releases 是否有新版本可用
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 检查是否有新版本
      tags:
      - Updater
...
  /api/v1/updater/update:
    post:
      consumes:
      - application/json
      description: 下载并安装新版本（公开接口，建议在生产环境中添加权限控制）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 执行更新
      tags:
      - Updater
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"